update grid layout

+++ src/widgets/layout.rs
@@ -1,20 +1,20 @@
 use crate::*;
 
-pub const DIR_LTR: u32 = 0;
-pub const DIR_RTL: u32 = 1;
-pub const DIR_DOWN: u32 = 2;
-pub const DIR_UP: u32 = 3;
+pub const DIR_LTR: usize = 0;
+pub const DIR_RTL: usize = 1;
+pub const DIR_DOWN: usize = 2;
+pub const DIR_UP: usize = 3;
 
 pub trait Layout {
     fn layout(node_idx: usize, origin: Point, glyph_size: Size) -> Point;
 }
 
-pub struct GridLayout<const DIR: u32, const WRAP: u32>;
+pub struct GridLayout<const DIR: usize, const WRAP: usize>;
 
-impl<const DIR: u32, const WRAP: u32> Layout for GridLayout<DIR, WRAP> {
+impl<const DIR: usize, const WRAP: usize> Layout for GridLayout<DIR, WRAP> {
     fn layout(node_idx: usize, origin: Point, size: Size) -> Point {
-        let idx = (node_idx as u32 % WRAP) as i32;
-        let wraps = (node_idx as u32 / WRAP) as i32;
+        let idx = (node_idx % WRAP) as i32;
+        let wraps = (node_idx / WRAP) as i32;
         let width = size.width as i32;
         let height = size.height as i32;
         let offset = match DIR {
